+++ src/commands/index.ts
@@ -3,6 +3,7 @@ export * from './controllers/files/component';
 export * from './controllers/files/layout';
 export * from './controllers/files/loading';
 export * from './controllers/files/page';
+export * from './controllers/terminal/prisma';
 export * from './controllers/terminal/project';
 export * from './controllers/terminal/serve';
 export * from './utils/functions';