+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-12-17
+
+### Added
+
+- Add Prisma snippets
+- Add Prisma commands
+
+### Changed
+
+- Update React Snippets
+
 ## [0.3.0] - 2023-12-16
 
 ### Added
@@ -23,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-nextjs-generator/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-nextjs-generator/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/ManuelGil/vscode-nextjs-generator/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/ManuelGil/vscode-nextjs-generator/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ManuelGil/vscode-nextjs-generator/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ManuelGil/vscode-nextjs-generator/releases/tag/v0.1.0